restructure for options and observer
