+++ lib/src/lib.rs
@@ -3,4 +3,6 @@
 pub mod client;
 mod crypto;
 pub mod models;
+pub mod observer;
+pub mod options;
 mod web;
